move resource_base
