+++ src/actinia_core/rest/location_management.py
@@ -44,7 +44,7 @@ from actinia_core.rest.base.user_auth import check_user_permissions
 from actinia_core.models.response_models import ProcessingResponseModel
 from actinia_core.models.response_models import SimpleResponseModel
 from actinia_core.models.response_models import MapsetInfoResponseModel
-from actinia_core.rest.resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.rest.persistent_processing import PersistentProcessing
 from actinia_core.rest.mapset_management import PersistentGetProjectionRegionInfo
 from actinia_core.core.common.redis_interface import enqueue_job
