+++ src/actinia_core/rest/map_layer_management.py
@@ -28,7 +28,7 @@ from flask import jsonify, make_response
 from flask_restful_swagger_2 import swagger
 import pickle
 from actinia_core.rest.persistent_processing import PersistentProcessing
-from actinia_core.rest.resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.request_parser import glist_parser, \
      extract_glist_parameters
