+++ src/actinia_core/rest/persistent_mapset_merger.py
@@ -28,7 +28,7 @@ import pickle
 from flask import jsonify, make_response
 
 from actinia_core.rest.persistent_processing import PersistentProcessing
-from actinia_core.rest.resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError, \
      AsyncProcessTermination
