+++ src/actinia_core/rest/raster_legend.py
@@ -28,7 +28,7 @@ Raster map renderer
 from flask import jsonify, make_response, Response
 from flask_restful import reqparse
 from .ephemeral_processing import EphemeralProcessing
-from .resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from tempfile import NamedTemporaryFile
 import os
