+++ src/actinia_core/rest/strds_raster_management.py
@@ -35,7 +35,7 @@ from actinia_api.swagger2.actinia_core.schemas.strds_raster_management import \
 
 from actinia_core.core.request_parser import where_parser
 from .persistent_processing import PersistentProcessing
-from .resource_base import ResourceBase
+from actinia_core.rest.base.resource_base import ResourceBase
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.core.common.exceptions import AsyncProcessError
 from actinia_core.models.response_models import \
